create AbstrGemaModel and ShotType

+++ homework4try2/src/main/java/com/a1arick/spbsu/homework4/try2/server/Point.java
@@ -3,19 +3,19 @@ package com.a1arick.spbsu.homework4.try2.server;
 import java.util.Objects;
 
 public class Point implements Comparable<Point>{
-    private final int x;
-    private final int y;
+    private final double x;
+    private final double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
@@ -35,6 +35,6 @@ public class Point implements Comparable<Point>{
 
     @Override
     public int compareTo(Point o) {
-        return Integer.compare(x, o.getX());
+        return Double.compare(x, o.getX());
     }
 }
